pyproject.toml
[project]
name = "memoraweave"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "langchain>=1.2.15",
    "langchain-google-genai>=4.2.1",
    "langchain-google-vertexai>=3.2.2",
    "langgraph>=1.1.6",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.44.0",
]
